requirements.txt
langchain==0.3.27
langchain-google-genai==2.0.10
langchain-ollama==0.3.6
langchain-core==0.3.75
langgraph==0.6.5
langgraph-checkpoint-sqlite==2.0.11
langmem==0.0.29
python-dotenv==1.1.1
PyPDF2==3.0.1
python-docx==1.2.0
pandas==2.3.1
openpyxl==3.1.5
langchain-text-splitters==0.3.9
langchain-chroma==0.2.5
aiofiles==24.1.0
aiosqlite==0.21.0
Crawl4AI==0.7.4
numpy==1.26.4
docling==2.44.0
mlx-vlm==0.3.3
torch==2.2.2
rich==13.9.4
pyperclip==1.9.0
ddgs<9.5.0
tavily-python


